LiquidAI:LFM2-VL-450M/requirements.txt
torch 
torchvision 
torchaudio
transformers 
accelerate 
huggingface_hub 
pillow
streamlit
uvicorn 
fastapi